Only deploy builds to Github Pages when pushed on master branch

+++ extra/travis_deploy_ghpages.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -ev
-if [ "${TRAVIS_PULL_REQUEST}" = "false" ]; then
+if [ "${TRAVIS_PULL_REQUEST}" = "false" ] && [ "${TRAVIS_BRANCH}" = "master" ]; then
   rm -rf out || exit 0;
   mkdir out;
   cd out
